Closes #1411: Remove play-service.json and add config to Firebase module
In this commit:
1. Move FirebaseHelper.init to a content provider so it gets init earlier.
2. Disable the content provider that Fabric and Firebase previously use.
3. Use Firebasehelper.bind() to flip the switch.
4. Remove gms plugin so no need to verify play-service.json.
5. Manually add the api key to resValue and generates at compile time
6. App module can fail back to firebase or firebase_no_op build type in fireabse module to inject the impl.
7. Add firebase build type to app module, so only people who interested
in Firebase need to use that build type.

TODOS
1. Make resValue accept environment variables.
2. Add the link build.gralde for firebase module, to denote the document
for manually create build id.
3. Add more document to let others know how Fabric and Firebase are
initialized and how to adjust it.
4. let firebase crashlytics accept debug mode
4. Test cases
  a. switch on/off Firebase, test RateApp Dialog. It should display
  default value when Firebase is off.
  b. make a crash, see if the crash apears in the backend ( may need to
  have debug mode first)
  c. switch on/off and test push notification : when app in the
  background / in the foreground / not running (6 test cases)
  d. in firebase console, see debug view and check for events.
