--- app/google-services.json
@@ -1,148 +1,0 @@
-{
-  "project_info": {
-    "project_number": "1062658243921",
-    "firebase_url": "https://remoteconfigsample-39b5e.firebaseio.com",
-    "project_id": "remoteconfigsample-39b5e",
-    "storage_bucket": "remoteconfigsample-39b5e.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:1062658243921:android:e11b6cbb185fee69",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket.debug.bitrise"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "1062658243921-pmiqk0hsrv01ir3vsr5sogdtbpf4alfh.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyCQWNEMAqJcG40XzLbIi0fcjzRvP1sUiBU"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 1,
-          "other_platform_oauth_client": []
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:1062658243921:android:72a93cf4c1451539",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket.debug.mozilla"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "1062658243921-pmiqk0hsrv01ir3vsr5sogdtbpf4alfh.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyCQWNEMAqJcG40XzLbIi0fcjzRvP1sUiBU"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 1,
-          "other_platform_oauth_client": []
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:1062658243921:android:12976494afaf30aa",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket.debug.nechen"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "1062658243921-2pjipqpatq45rv0kl77nr4i7rtu301bd.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "org.mozilla.rocket.debug.nechen",
-            "certificate_hash": "3b23f4306181d85e3c341111db7136e42d3a185f"
-          }
-        },
-        {
-          "client_id": "1062658243921-pmiqk0hsrv01ir3vsr5sogdtbpf4alfh.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyCQWNEMAqJcG40XzLbIi0fcjzRvP1sUiBU"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "1062658243921-pmiqk0hsrv01ir3vsr5sogdtbpf4alfh.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:1062658243921:android:5f7b2773da76a1cb",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket.debug.null"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "1062658243921-pmiqk0hsrv01ir3vsr5sogdtbpf4alfh.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyCQWNEMAqJcG40XzLbIi0fcjzRvP1sUiBU"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 1,
-          "other_platform_oauth_client": []
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
