--- app/src/beta/google-services.json
@@ -1,166 +1,0 @@
-{
-  "project_info": {
-    "project_number": "994508393765",
-    "firebase_url": "https://zerda-dcf76.firebaseio.com",
-    "project_id": "zerda-dcf76",
-    "storage_bucket": "zerda-dcf76.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:994508393765:android:b4171814c52b62df",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket.debug.maliu"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "994508393765-fekfivvc4gb30nh3fcf7ouac03g3qltt.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "org.mozilla.rocket.debug.maliu",
-            "certificate_hash": "45c940c6d0ea54fb9f3f423ca068df8f7697782d"
-          }
-        },
-        {
-          "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBpew1lRW45fWwQhhhKUz_ZGSJjZQ7kSRo"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:994508393765:android:68d62f85027bc145",
-        "android_client_info": {
-          "package_name": "gro.allizom.zelda.beta"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "994508393765-2d7aq5mlnhafso1onqcqi908ekdvbcfu.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "gro.allizom.zelda.beta",
-            "certificate_hash": "8956c0d3a7ace284a2566f1806c0d0019d3428d2"
-          }
-        },
-        {
-          "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBpew1lRW45fWwQhhhKUz_ZGSJjZQ7kSRo"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:994508393765:android:06dcf70b6f8303ce",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBpew1lRW45fWwQhhhKUz_ZGSJjZQ7kSRo"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 1,
-          "other_platform_oauth_client": []
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:994508393765:android:993448349fb747a5",
-        "android_client_info": {
-          "package_name": "org.mozilla.rocket.coverage"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBpew1lRW45fWwQhhhKUz_ZGSJjZQ7kSRo"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "994508393765-r766o21k3u0pmjg509vnku0jjqo00b6a.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
