+++ app/src/main/res/raw/third_party_license_metadata
@@ -1,83 +1,62 @@
 0:53 checkstyle
 55:33 antlr
-90:47 firebase-core
-139:47 firebase-measurement-connector-impl
-188:47 firebase-analytics
-237:47 firebase-analytics-impl
-286:47 firebase-iid
-335:47 play-services-stats
-384:47 firebase-common
-433:47 firebase-measurement-connector
-482:47 play-services-ads-identifier
-531:47 firebase-iid-interop
-580:47 play-services-base
-629:47 play-services-tasks
-678:47 play-services-basement
-727:47 play-services-measurement-base
-776:46 support-v4
-824:46 support-media-compat
-872:46 support-fragment
-920:46 support-core-utils
-968:46 support-core-ui
-1016:46 support-compat
-1064:46 support-annotations
-1112:46 runtime
-3678:46 common
-1256:37 findbugs
-4286:46 jsr305
-1343:46 bcel-findbugs
-1391:37 jFormatString
-1430:158 AppleJavaExtensions
-1590:38 jaxen
-1630:46 xml-apis
-1678:46 kotlin-annotation-processing-gradle
-1820:46 compiler
-1868:46 kotlin-compiler-embeddable
-1916:46 migration
-4238:46 kotlin-stdlib
-2012:46 javapoet
-3971:46 annotations
-2250:46 sqlite-jdbc
-2298:46 kotlin-script-runtime
-2394:46 lint-gradle
-2442:46 builder
-2490:46 lint
-2538:46 lint-checks
-2586:46 lint-api
-2634:46 manifest-merger
-2682:46 sdk-common
-2730:46 lint-gradle-api
-2778:46 gradle-api
-2826:46 builder-model
-2874:46 lint-kotlin
-2922:46 intellij-core
-2970:46 uast
-3018:46 groovy-all
-3066:46 kotlin-reflect
-3114:46 sdklib
-3162:46 repository
-3210:46 kotlin-stdlib-jre8
-3258:46 builder-test-api
-3306:46 ddmlib
-3354:40 bcpkix-jdk15on
-3396:40 bcprov-jdk15on
-3438:46 tracker
-3486:46 shared
-3534:46 protos
-3582:46 layoutlib-api
-3630:46 dvlib
-3726:46 apksig
-3774:46 javawriter
-3822:47 fastutil
-3871:50 jopt-simple
-3923:46 json-simple
-4019:121 kxml2
-4142:46 kotlin-compiler
-4190:46 kotlin-stdlib-jre7
-4334:46 j2objc-annotations
-4382:55 stax-ex
-4439:102 jaxb-api
-4543:51 javacc
-4596:38 saxon
-4636:46 jcommander
-4684:46 gson
+90:37 findbugs
+3168:46 jsr305
+177:46 bcel-findbugs
+225:37 jFormatString
+264:158 AppleJavaExtensions
+424:38 jaxen
+464:46 xml-apis
+512:46 kotlin-annotation-processing-gradle
+654:46 compiler
+702:46 kotlin-compiler-embeddable
+750:46 migration
+3120:46 kotlin-stdlib
+846:46 javapoet
+2853:46 annotations
+2560:46 common
+1084:46 sqlite-jdbc
+1132:46 kotlin-script-runtime
+1180:46 support-annotations
+1276:46 lint-gradle
+1324:46 builder
+1372:46 lint
+1420:46 lint-checks
+1468:46 lint-api
+1516:46 manifest-merger
+1564:46 sdk-common
+1612:46 lint-gradle-api
+1660:46 gradle-api
+1708:46 builder-model
+1756:46 lint-kotlin
+1804:46 intellij-core
+1852:46 uast
+1900:46 groovy-all
+1948:46 kotlin-reflect
+1996:46 sdklib
+2044:46 repository
+2092:46 kotlin-stdlib-jre8
+2140:46 builder-test-api
+2188:46 ddmlib
+2236:40 bcpkix-jdk15on
+2278:40 bcprov-jdk15on
+2320:46 tracker
+2368:46 shared
+2416:46 protos
+2464:46 layoutlib-api
+2512:46 dvlib
+2608:46 apksig
+2656:46 javawriter
+2704:47 fastutil
+2753:50 jopt-simple
+2805:46 json-simple
+2901:121 kxml2
+3024:46 kotlin-compiler
+3072:46 kotlin-stdlib-jre7
+3216:46 j2objc-annotations
+3264:55 stax-ex
+3321:102 jaxb-api
+3425:51 javacc
+3478:38 saxon
+3518:46 jcommander
+3566:46 gson
