+++ app/src/main/res/raw/third_party_licenses
@@ -1,29 +1,5 @@
 http://www.gnu.org/licenses/old-licenses/lgpl-2.1.txt
 http://www.antlr.org/license.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-https://developer.android.com/studio/terms.html
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
-http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.gnu.org/licenses/lgpl.html
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
@@ -40,6 +16,7 @@ http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.opensource.org/licenses/bsd-licensehttp://www.apache.org/licenses/LICENSE-2.0.txt
+http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
 http://www.apache.org/licenses/LICENSE-2.0.txt
