+++ firebase/src/main/AndroidManifest.xml
@@ -1,7 +1,129 @@
+<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.firebase">
+    xmlns:tools="http://schemas.android.com/tools"
+    package="org.mozilla.firebase">
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.WAKE_LOCK" />
+
+    <application>
+
+        <!--<service-->
+            <!--android:name="com.google.firebase.messaging.FirebaseMessagingService"-->
+            <!--android:exported="true">-->
+
+            <!--<intent-filter android:priority="-500">-->
+
+                <!--<action android:name="com.google.firebase.MESSAGING_EVENT" />-->
+            <!--</intent-filter>-->
+        <!--</service>-->
+
+
+        <!--<receiver-->
+            <!--android:name="com.google.android.gms.analytics.AnalyticsReceiver"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="false" />-->
+
+        <!--<service-->
+            <!--android:name="com.google.android.gms.analytics.AnalyticsService"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="false" />-->
+
+        <!--<service-->
+            <!--android:name="com.google.android.gms.analytics.AnalyticsJobService"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="false"-->
+            <!--android:permission="android.permission.BIND_JOB_SERVICE" />-->
+
+        <!--<service android:name="com.google.firebase.components.ComponentDiscoveryService">-->
+
+            <!--<meta-data-->
+                <!--android:name="com.google.firebase.components:com.google.firebase.analytics.connector.internal.AnalyticsConnectorRegistrar"-->
+                <!--android:value="com.google.firebase.components.ComponentRegistrar" />-->
+
+            <!--<meta-data-->
+                <!--android:name="com.google.firebase.components:com.google.firebase.iid.Registrar"-->
+                <!--android:value="com.google.firebase.components.ComponentRegistrar" />-->
+        <!--</service>-->
+
+        <!--<receiver-->
+            <!--android:name="com.google.android.gms.measurement.AppMeasurementReceiver"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="false" />-->
+
+        <!--<receiver-->
+            <!--android:name="com.google.android.gms.measurement.AppMeasurementInstallReferrerReceiver"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="true"-->
+            <!--android:permission="android.permission.INSTALL_PACKAGES">-->
+
+            <!--<intent-filter>-->
+
+                <!--<action android:name="com.android.vending.INSTALL_REFERRER" />-->
+            <!--</intent-filter>-->
+        <!--</receiver>-->
+
+        <!--<service-->
+            <!--android:name="com.google.android.gms.measurement.AppMeasurementService"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="false" />-->
+
+        <!--<service-->
+            <!--android:name="com.google.android.gms.measurement.AppMeasurementJobService"-->
+            <!--android:enabled="true"-->
+            <!--android:exported="false"-->
+            <!--android:permission="android.permission.BIND_JOB_SERVICE" />-->
+
+        <!--<receiver-->
+            <!--android:name="com.google.firebase.iid.FirebaseInstanceIdReceiver"-->
+            <!--android:exported="true"-->
+            <!--android:permission="com.google.android.c2dm.permission.SEND">-->
+
+            <!--<intent-filter>-->
+
+                <!--<action android:name="com.google.android.c2dm.intent.RECEIVE" />-->
+
+                <!--<category android:name="org.mozilla.rocket.debug.mozilla" />-->
+            <!--</intent-filter>-->
+        <!--</receiver>-->
+
+        <!--<service-->
+            <!--android:name="com.google.firebase.iid.FirebaseInstanceIdService"-->
+            <!--android:exported="true">-->
+
+            <!--<intent-filter android:priority="-500">-->
+
+                <!--<action android:name="com.google.firebase.INSTANCE_ID_EVENT" />-->
+            <!--</intent-filter>-->
+        <!--</service>-->
+
+        <!--<activity-->
+            <!--android:name="com.google.android.gms.common.api.GoogleApiActivity"-->
+            <!--android:exported="false"-->
+            <!--android:theme="@android:style/Theme.Translucent.NoTitleBar" />-->
+
+        <provider
+            android:name="com.google.firebase.provider.FirebaseInitProvider"
+            android:authorities="${applicationId}.firebaseinitprovider"
+            android:enabled="false"
+            android:exported="false"
+            android:initOrder="100"
+            tools:node="remove" />
+
+        <!--<meta-data-->
+            <!--android:name="com.google.android.gms.version"-->
+            <!--android:value="@integer/google_play_services_version" />-->
+
+        <!--<meta-data android:name="firebase_crashlytics_collection_enabled" android:value="@bool/enableCrashlytics" />-->
+
+
+        <provider
+            android:name="com.crashlytics.android.CrashlyticsInitProvider"
+            android:authorities="${applicationId}.crashlyticsinitprovider"
+            android:enabled="false"
+            android:exported="false"
+            android:initOrder="100"
+            tools:node="remove" />
+    </application>
 </manifest>
